update human filesize display
