+++ src/PenguinUpload/ClientApp/src/stylus/main.styl
@@ -15,6 +15,6 @@ $theme := {
 
 /* Overrides */
 
-h2, h3, h4, h5, h6 {
+h2.light, h3.light, h4.light, h5.light, h6.light {
   font-weight: 300;
 }
